Add a reduction to the "buggy" (bug-compatible) non-dominated sorting.
The term "bug-compatible" refers to the (hypothetically) original
implementation of non-dominated sorting by Deb, which treated equal
points as dominating one another. As a result, there was a complete
mess when equal points existed in the input. In general (but not always)
equal point received sequentially increasing ranks.

This kind of things can sometimes be desired, so that whenever we really
have many different individuals with equal phenotypes, those who are
dominated by them become much worse. A normally looking extension
to that can be introduced: when we have a group of equal points,
the first one receives the maximum rank among dominators + 1,
the second one gets this value plus 1, the third one gets one more
and so on.

To prevent bias due to the order the points are given in,
the particular order between these points must be determined at random.
However, for testing purposes, we always think that the point which goes
first in the input is better. This is easily fixable, however.